Add ledger encryption
Since the content of the ledger is sensitive, it is possible to encrypt
the file. The system will decrypt it before reading it (or performing
any other operation) and encrypt it back immediately. This feature can
be turned on/off through ENCRYPTION variable and the credentials can
also be read from environment variables or by manual input.